+++ local/testing/api-tests/tavern_helpers.py
@@ -3,6 +3,7 @@ from datetime import datetime
 
 import requests
 import os
+from box import Box
 
 
 def supplyChainImpacted_is_Yes(response):
@@ -59,6 +60,26 @@ def ESS_job_parameter_are_as_requested(response):
     aspects_list = parameter.get("aspects")
     assert 'PartSiteInformationAsPlanned' in aspects_list
     assert 'PartAsPlanned' in aspects_list
+
+def tombstone_for_EssValidation_are_correct(response, expectedTombstone):
+    error_list = response.json().get("tombstones")
+
+    for i in error_list:
+        print("Given tombstone: ", i)
+        catenaXId = i.get("catenaXId")
+        print("Given catenaXID: ", catenaXId)
+        processingErrorStep = i.get("processingError").get("processStep")
+        print("Processstep in ProcessingError: ", processingErrorStep)
+        processingErrorDetail = i.get("processingError").get("errorDetail")
+        print("ErrorMessage: ", processingErrorDetail)
+        processingErrorLastAttempt = i.get("processingError").get("lastAttempt")
+        print("LastAttempt: ", processingErrorLastAttempt)
+        processingErrorRetryCounter = i.get("processingError").get("retryCounter")
+        print("RetryCounter: ", processingErrorRetryCounter)
+        assert 'EssValidation' in processingErrorStep
+        assert expectedTombstone in processingErrorDetail
+        assert processingErrorLastAttempt is not None
+        assert 0 is processingErrorRetryCounter
 
 
 ############################## /\ ESS helpers /\ ##############################
@@ -287,6 +308,20 @@ def order_informations_for_batchprocessing_are_given(response, amount_batches):
         assert ('https://irs.dev.demo.catena-x.net/irs/orders' in batches.get("batchUrl")) or ('https://irs.int.demo.catena-x.net/irs/orders' in batches.get("batchUrl"))
         assert batches.get("batchProcessingState") == 'INITIALIZED'
         assert batches.get("errors") is None
+
+
+def getBatchId(response, batchId_number):
+    batches_list = response.json().get("batches")
+    batchId_list = []
+    for batches in batches_list:
+        batchId_list.append(batches.get("batchId"))
+    return Box({"batch_id": batchId_list[batchId_number]})
+
+
+def check_batches_are_canceled_correctly(response):
+    jobs_list = response.json().get("jobs")
+    for jobs in jobs_list:
+        assert jobs.get("state") == 'CANCELED'
 
 
 def job_parameter_are_as_requested(response):