+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- ESS
+  - Added 'hops' parameter to Supply Chain Aspect model - contains relative distance in the supply chain
 
 ## [4.0.0] - 2023-10-27
 ### Added