+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - ESS
-  - Added 'hops' parameter to Supply Chain Aspect model - contains relative distance in the supply chain
+  - Added 'hops' parameter to SupplyChainImpacted Aspect model - contains relative distance in the supply chain
 
 ## [4.0.0] - 2023-10-27
 ### Added